[KUNLUNXIN] fix aminmax, apply_repetition_penalties, leaky_relu, leaky_relu_, leaky_relu_out, mm, weight_norm accuracy tests

+++ src/flag_gems/runtime/backend/_kunlunxin/ops/apply_repetition_penalties.py
@@ -34,8 +34,9 @@ def _repetition_penalty_kernel(
 
     prompt_mask = tl.load(prompt_mask_ptr + mask_idx, mask=valid_vocab, other=False)
     output_mask = tl.load(output_mask_ptr + mask_idx, mask=valid_vocab, other=False)
-    logits = tl.load(logits_ptr + logits_idx, mask=valid_vocab, other=0.0)
+    logits = tl.load(logits_ptr + logits_idx, mask=valid_vocab, other=0.0).to(tl.float32)
 
+    penalty = penalty.to(tl.float32)
     is_repeated = prompt_mask | output_mask
 
     logits = tl.where(is_repeated & (logits > 0), logits / penalty, logits)
